Move codepage functionality to os2 module
