+++ crates/fonticulus/src/utils.rs
@@ -1,11 +1,3 @@
-pub fn int_list_to_num(int_list: &[u8]) -> u32 {
-    let mut flags = 0;
-    for flag in int_list {
-        flags |= 1 << flag;
-    }
-    flags
-}
-
 pub fn adjust_offset<T>(offset: T, angle: f64) -> i32
 where
     T: Into<f32>,
